pyproject.toml
[project]
name = "vector-databases-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.25",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langchain-postgres>=0.0.17",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.30.0",
    "pgvector>=0.3.6",
    "pinecone>=7.3.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
]
